refactor: extract LazyComponent
