+++ src/index.tsx
@@ -7,47 +7,13 @@ import React, {
   useState,
 } from 'react';
 import {
+  type NativeSyntheticEvent,
+  type NativeScrollEvent,
   FlatList,
   useWindowDimensions,
   View,
-  ActivityIndicator,
 } from 'react-native';
-import {
-  type NativeSyntheticEvent,
-  type NativeScrollEvent,
-} from 'react-native';
-
-interface LazyComponentProps {
-  componentKey: number;
-  currentKey: number;
-  component: React.ReactNode;
-}
-const LazyComponent = (props: LazyComponentProps) => {
-  const { componentKey, currentKey, component } = props;
-
-  const [hasRendered, setHasRendered] = useState(false);
-
-  const { width: windowWidth } = useWindowDimensions();
-
-  useEffect(() => {
-    if (!hasRendered && currentKey === componentKey) {
-      setHasRendered(true);
-    }
-  }, [currentKey, componentKey, hasRendered]);
-
-  if (hasRendered) return component;
-  return (
-    <View
-      style={{
-        width: windowWidth,
-        alignItems: 'center',
-        justifyContent: 'center',
-      }}
-    >
-      <ActivityIndicator />
-    </View>
-  );
-};
+import { LazyComponent } from './LazyComponent';
 
 interface PagerViewProps {
   initialPage?: number;
